Add untested first attempt to include Kubernetes Namespace labels and annotations
At the moment, the in-memory cache of Namespace Metadata is never refreshed, and never purged.

There is no configuration to opt-out of Namespace metadata being added to events aside from the existing Pod-level labels and annotations filters.

It also introduces a breaking change to the metadata added to the event by converting namespace from a string containing the name to an object with a name member and also labels and annotations members.